Nerfs ahoy
Nerfed Sword / Shield of Darkness - now comparable to a high-tuned 1-cost, but lower than any 2-cost.

Nerfed Phantom Fireball - not that difficult to make it a 24 damage 2-cost and scales 3x with strength.  Now hits once for 9 and once again for 9 if enemy is burning.

Reworded Instant Inferno to fit better on the card template.