Colorless cards migration & rebuild - potential breakage
The Study the Spire mechanic and related cards/powers are now in a new dedicated 'Expansion Content' package - this should include all the content that is added to even the vanilla game.  Former Study cards are in process of being rebuilt into the new code templates, getting rid of all the old now-unused functions and streamlining for consistency with Hexa's code style.  These are very much a WIP and are disabled until fully functional.  This shouldn't have any impact on Slimebound as the Study cards are no longer in its kit, but has some breakage risks.

Slimebound Mod restructured to use the auto-add cards function, not a manual list.

Resources for colorless cards migrated to ExpansionContent package