Framework updates to the Char Boss system
Since the 'run simulator' now handles card upgrades, GetPriority for upgrades has been removed.  GetPriority now refers to the priority of the card when the boss decides the hand order.

GetPriority overrides added to more cards to update priority under different circumstances.

Added new gameplay tags for Setup and Attack, applying them to all Ironclad cards as examples.

Rather than drawing a card at random, the boss will draw specific cards to make turns a bit more fair.   All bosses start with a Setup turn, then follow with an Attack turn, then repeat.

A Setup turn draws one Setup card, then tries to draw the remaining cards for the hand as cards that are not Setup or Attack.

A Attack turn draws two Attack cards, then tries to draw the remaining cards for the hand as cards that are not Setup or Attack.

In either case, if non-Setup, non-Attack cards are available, the priorities will keep shfiting until a suitable card is found.

A new boolean for ForceDraw is added for status/curse cards, and is applied to whatever card Headbutt places on top of the deck.  This indicates to the draw logic to not override and try to draw another card, to instead draw that card as it lies regardless of phase.