Fix rare crash with Blank Card, Gremlin Dance, and starter relic
The crash occurs when the Gremlin starter relic swaps to the Mad
gremlin after the card is already in queue. Since the card in queue is
expecting to find a different gremlin in front, it crashes on Mad's
multi-damage attack because the multiDamage field is null. This isn't
obvious because the stack trace has incorrect line numbers within code
patched by basemod.

The fix of calling applyPowers was inspired by the code of the vanilla
card Havoc.

Rather than trying to find a different fix inside Gremlin Dance, this
fix ensures Blank Card will call applyPowers just before playing the
card, which hopefully avoids similar corner case crashes with
start-of-battle relics and other cards.