Minor fixes.
Recursive Strike now has a Strike Tag.
Rain of Embers secondary effect works with Chemical X.
Seals now have a visible upgrade button in card library.

Fixed to a typo in Technical Jig's description.

Changed the descriptions of Finish Him and Burning Touch to be less confusing.

Rest of this is just a tangent.
 This description is worded that way to suggest that Hp check happens before dealing damage, so every fight does not require math similar to Slimeboss fight, so please do not change it unless the card is reworked. Also I am keeping the percent as something that can be changed in the card because I feel that it would be required to balance the card later.